% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typography.R
\name{clean_decimal}
\alias{clean_decimal}
\title{Clean decimal formatting}
\usage{
clean_decimal(x)
}
\arguments{
\item{x}{numeric value to format}
}
\value{
Character string with trailing zeros and decimal point removed
}
\description{
Remove trailing zeros and decimal points from numeric values formatted as
strings. Used internally to format point values in pica notation.
}
\examples{
clean_decimal(2.3)
clean_decimal(2.0)
clean_decimal(2.30000000000002)
}
\keyword{internal}
