% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_preview.R
\name{color_preview}
\alias{color_preview}
\title{Preview a list of hex colors}
\usage{
color_preview(x, name = NULL, ...)
}
\arguments{
\item{x}{character vector of colors in hex format}

\item{name}{string of name to print in the palette; if not given, the
function looks for a name attribute in the object}

\item{...}{does nothing for now}
}
\description{
Show a plot of colors. Code adapted from Karthik Ram's wesandersonr package
(https://github.com/karthik/wesanderson/blob/master/R/colors.R).
}
\examples{
\dontrun{
color_preview(c("#0074D9", "#2ECC40", "#FFDC00", "#FF4136"))

color_preview(c("#0074D9", "#2ECC40", "#FFDC00", "#FF4136"),
              name = "Some colors")

color_preview(viridisLite::viridis(6, begin = 0.15, end = 0.85,
                                   option = "inferno"))
}
}
