% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typography.R
\name{picas_to_in}
\alias{picas_to_in}
\title{Convert picas to inches}
\usage{
picas_to_in(picas)
}
\arguments{
\item{picas}{Numeric value in picas, or character string in InDesign format
(e.g., "1p6" for 1 pica and 6 points)}
}
\value{
Numeric value in inches
}
\description{
Convert picas to inches
}
\examples{
picas_to_in(1.5)
picas_to_in("1p6")
}
