% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typography.R
\name{picas_to_num}
\alias{picas_to_num}
\title{Convert picas to numeric}
\usage{
picas_to_num(picas)
}
\arguments{
\item{picas}{numeric value in picas, or character string in InDesign format
(e.g., "1p6" for 1 pica and 6 points)}
}
\value{
Numeric value in picas
}
\description{
Convert picas to numeric
}
\examples{
picas_to_num(1.5)
picas_to_num("1p6")
picas_to_num("1p")
}
\keyword{internal}
