% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typography.R
\name{picas_to_pts}
\alias{picas_to_pts}
\title{Convert picas to points}
\usage{
picas_to_pts(picas)
}
\arguments{
\item{picas}{numeric value in picas, or character string in InDesign format
(e.g., "1p6" for 1 pica and 6 points)}
}
\value{
Numeric value in points (72 points = 1 inch)
}
\description{
Convert picas to points
}
\examples{
picas_to_pts(1.5)
picas_to_pts("1p6")
}
