% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rath-package.r
\docType{package}
\name{rath-package}
\alias{rath}
\alias{rath-package}
\title{rath: Andrew Heiss's custom R code}
\description{
Provides a library of miscellaneous R functions.
}
\author{
\strong{Maintainer}: Andrew Heiss \email{andrew@andrewheiss.com} (\href{https://orcid.org/0000-0002-3948-3914}{ORCID})

}
\keyword{internal}
