% Generated by roxygen2 (4.0.0.99): do not edit by hand
\name{theme_ath}
\alias{theme_ath}
\title{Custom ggplot2 theme}
\usage{
theme_ath(base_size = 12)
}
\arguments{
\item{base_size}{Base font size (defaults to 12)}
}
\description{
Provides a modified version of \code{theme_bw()}.
}
\examples{
library(ggplot2)
ggplot(mtcars, aes(x=mpg, y=wt, colour=as.factor(cyl))) +
  geom_point() + theme_ath()
}

