% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/theme_geektool.R
\name{theme_geektool}
\alias{theme_geektool}
\title{theme_geektool}
\usage{
theme_geektool()
}
\description{
A marginless blank theme for showing graphs on a dark background
}
\examples{
library(ggplot2)
ggplot(mtcars, aes(x=mpg, y=wt, colour=as.factor(cyl))) +
  geom_point() + theme_ath()
}

