% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/theme_ath.R
\name{theme_ath}
\alias{theme_ath}
\title{My custom ggplot theme}
\usage{
theme_ath(base_size = 12, base_family = "", legend_bottom = FALSE,
  margin_bottom = -0.5)
}
\arguments{
\item{base_size}{base font size (defaults to 12)}

\item{base_family}{base font family}

\item{legend_bottom}{set this to TRUE to place the legend at the bottom of the plot}

\item{margin_bottom}{bottom plot margin, in lines (\bold{only used when} \code{legend_bottom = TRUE}) (defaults to -0.5)}
}
\description{
Provides a modified version of \code{\link{theme_bw}}.
}
\examples{
library(ggplot2)
ggplot(mtcars, aes(x=mpg, y=wt, colour=as.factor(cyl))) +
  geom_point() + theme_ath()

ggplot(mtcars, aes(x=mpg, y=wt, colour=as.factor(cyl))) +
  geom_point() + theme_ath(legend_bottom=TRUE)
}

